[spellbook] Remove obsolete spell creation and testing files
Delete `fn.test.ts` and `fn.ts` as they are no longer needed in the codebase, streamlining the spellbook functionality and reducing clutter.
